Move all popups into the popups folder
