+++ client/src/popups/EventPopup.scss
