+++ client/src/popups/Popup.scss
